// src/apiV2/a7-service/model/userLoginDto.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * A7 Backend API
 * API документация для A7 Backend - платформа для фотографов
 * OpenAPI spec version: 1.0.0
 */

export interface UserLoginDto {
  /** Email пользователя */
  email: string;
  /** Пароль пользователя */
  password: string;
}